Move proxied containers to commons
To get rid of toxiproxy deprecation in classic spring as well, and to
get rid of code duplication.

Sharing test code follows the default gradle way, as documented in
[Testing Java / using test
fixtures](https://docs.gradle.org/current/userguide/java_testing.html#sec:java_test_fixtures).
