+++ outbox-kafka-spring-reactive/build.gradle.kts
@@ -9,7 +9,6 @@ dependencies {
     val springVersion = "6.1.2"
     val springDataVersion = "3.2.1"
     val kafkaVersion = "3.6.1"
-    val springKafkaVersion = "3.0.12"
     val testcontainersVersion = "1.19.3"
     val log4jVersion = "2.22.0"
 
@@ -22,13 +21,13 @@ dependencies {
     implementation("org.apache.kafka:kafka-clients:$kafkaVersion")
     "protobufSupportImplementation"("com.google.protobuf:protobuf-java:${rootProject.extra["protobufVersion"]}")
     implementation("com.fasterxml.jackson.core:jackson-databind:2.16.1")
-    implementation(project(":commons"))
     implementation("org.slf4j:slf4j-api:2.0.9")
     implementation("javax.annotation:javax.annotation-api:1.3.2")
+    implementation(project(":commons"))
 
     // testing
+    testImplementation(testFixtures(project(":commons")))
     testImplementation("org.springframework.boot:spring-boot-autoconfigure:3.2.1")
-    testImplementation("org.springframework:spring-test:$springVersion")
     testImplementation("io.projectreactor:reactor-test:3.6.1")
 
     testRuntimeOnly("org.junit.platform:junit-platform-launcher")
@@ -39,16 +38,11 @@ dependencies {
     testImplementation("org.testcontainers:junit-jupiter:$testcontainersVersion")
     testImplementation("org.springframework:spring-jdbc:$springVersion")
     testImplementation("io.r2dbc:r2dbc-pool:1.0.1.RELEASE")
-    testImplementation("org.testcontainers:postgresql:$testcontainersVersion")
-    testImplementation("org.testcontainers:kafka:$testcontainersVersion")
-    testImplementation("org.testcontainers:toxiproxy:$testcontainersVersion")
     // update gson version to fix a conflict of toxiproxy dependency and spring GsonAutoConfiguration
     testRuntimeOnly("com.google.code.gson:gson:2.10.1")
     testImplementation("org.postgresql:postgresql:42.7.1")
     testImplementation("org.flywaydb:flyway-database-postgresql:10.4.1")
     testImplementation("org.flywaydb.flyway-test-extensions:flyway-spring-test:9.5.0")
-    testImplementation("org.springframework.kafka:spring-kafka:$springKafkaVersion")
-    testImplementation("org.springframework.kafka:spring-kafka-test:$springKafkaVersion")
     testImplementation("org.awaitility:awaitility:4.2.0")
     testImplementation("org.apache.logging.log4j:log4j-core:$log4jVersion")
     testImplementation("org.apache.logging.log4j:log4j-slf4j2-impl:$log4jVersion")
