+++ outbox-kafka-spring/src/test/java/one/tomorrow/transactionaloutbox/service/OutboxUsageIntegrationTest.java
@@ -17,7 +17,7 @@ package one.tomorrow.transactionaloutbox.service;
 
 import one.tomorrow.transactionaloutbox.IntegrationTestConfig;
 import one.tomorrow.transactionaloutbox.KafkaTestSupport;
-import one.tomorrow.transactionaloutbox.ProxiedKafkaContainer;
+import one.tomorrow.transactionaloutbox.commons.ProxiedKafkaContainer;
 import one.tomorrow.transactionaloutbox.model.OutboxLock;
 import one.tomorrow.transactionaloutbox.model.OutboxRecord;
 import one.tomorrow.transactionaloutbox.repository.OutboxLockRepository;
@@ -48,9 +48,8 @@ import java.time.Duration;
 import java.util.List;
 
 import static one.tomorrow.transactionaloutbox.IntegrationTestConfig.DEFAULT_OUTBOX_LOCK_TIMEOUT;
-import static one.tomorrow.transactionaloutbox.KafkaTestSupport.createConsumer;
-import static one.tomorrow.transactionaloutbox.KafkaTestSupport.producerProps;
-import static one.tomorrow.transactionaloutbox.ProxiedKafkaContainer.bootstrapServers;
+import static one.tomorrow.transactionaloutbox.commons.CommonKafkaTestSupport.*;
+import static one.tomorrow.transactionaloutbox.commons.ProxiedKafkaContainer.bootstrapServers;
 import static one.tomorrow.transactionaloutbox.service.SampleService.Topics.topic1;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
